requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
base58==2.1.0
bleach==3.2.2
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fpdf==1.7.2
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaleido==0.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.13
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pyzmq==21.0.1
requests==2.25.1
retrying==1.3.3
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
streamlit==0.79.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
